+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+-   Fix bug where browser in result table was displayed as `[object Object]`
+    instead of its name.
+
 <!-- Add new, unreleased changes here. -->
 
 ## [0.4.8] 2019-07-08